requirements.txt
flask
prometheus-client
python-logging-loki

